requirements.txt
django
pillow
django-crispy-forms
django-cleanup
django-filter